bugfix: separate sorting logic and remove stale times on list of contests
