+++ packages/react-app-revamp/components/_pages/ListContests/Contest/index.tsx
@@ -103,6 +103,7 @@ const Contest: FC<ContestProps> = ({ contest, compact, loading, rewards, allowTo
         setSubmissionStatus("Submissions open in:");
       }
     } else if (now.isBefore(moment(contest.vote_start_at))) {
+      console.log({ contest });
       setSubmissionStatus("Submissions are open");
 
       const secondsLeft = moment(contest.vote_start_at).diff(now, "seconds");
