+++ packages/react-app-revamp/pages/contests/past/index.tsx
@@ -13,11 +13,6 @@ function useContests(initialData: any) {
   const [page, setPage] = useState(0);
   const { address } = useAccount();
 
-  const queryOptions = {
-    keepPreviousData: true,
-    staleTime: 5000,
-  };
-
   //@ts-ignore
   if (initialData?.data) queryOptions.initialData = initialData.data;
 
@@ -26,7 +21,7 @@ function useContests(initialData: any) {
     data: contestData,
     error,
     isFetching: isContestDataFetching,
-  } = useQuery(["pastContests", page, address], () => getPastContests(page, ITEMS_PER_PAGE, address), queryOptions);
+  } = useQuery(["pastContests", page, address], () => getPastContests(page, ITEMS_PER_PAGE, address));
 
   const { data: rewardsData, isFetching: isRewardsFetching } = useQuery(
     ["rewards", contestData],
@@ -110,7 +105,10 @@ export async function getStaticProps() {
 
     const result = await supabase
       .from("contests_v3")
-      .select("created_at, start_at, end_at, address, author_address, network_name, vote_start_at, featured, title, type, summary, prompt", { count: "exact" })
+      .select(
+        "created_at, start_at, end_at, address, author_address, network_name, vote_start_at, featured, title, type, summary, prompt",
+        { count: "exact" },
+      )
       // all rows whose votes end date is < to the current date.
       .lt("end_at", new Date().toISOString())
       .order("end_at", { ascending: false })
